Add dev specifc dockerfiles
Although many such files already exists, we wanted to be able to do
everything from docker compose, meaning to build the necessary images,
run the local network, and also mount specific source code which we can
rebuild when needed for faster prototyping or experimentation. Also, we
want to simplify the Dockerfile such that rather than relying on 3
images, we now rely on one single Dockerfile, which can probably be
simplified as time goes.