Add make target to (only) build libgo_cosmwasm.so
Without building the trusted enclave code
(librust_cosmwasm_enclave.signed.so).

It's important to note that some (if not all) of the artifacts built as
part of building librust_cosmwasm_enclave are necessary dependencies to
build libgo_cosmwasm.so. Thus, the added target to only build
libgo_cosmwasm.so assumes that the required enclave dependencies have
been already built and are the expected locations on the filesystem.