(sbellem) squash commits
* Add log statements in go-cosmwasm rust code
* Log tx hash in secretd (keeper) go code
* Add cosmos-sdk as a git submodule
* Add IAS api key for testnet
* Add quick way to build artifacts
* Add helper docs for running tests

dockerfiles

* enclave-test.Dockerfile
* Add dev specifc dockerfiles

Although many such files already exists, we wanted to be able to do
everything from docker compose, meaning to build the necessary images,
run the local network, and also mount specific source code which we can
rebuild when needed for faster prototyping or experimentation. Also, we
want to simplify the Dockerfile such that rather than relying on 3
images, we now rely on one single Dockerfile, which can probably be
simplified as time goes.

* Update hacking/docker-compose.yml
* Add quick way to build untrusted artifacts
* Add make target to (only) build libgo_cosmwasm.so

* Without building the trusted enclave code
(librust_cosmwasm_enclave.signed.so).

It's important to note that some (if not all) of the artifacts built as
part of building librust_cosmwasm_enclave are necessary dependencies to
build libgo_cosmwasm.so. Thus, the added target to only build
libgo_cosmwasm.so assumes that the required enclave dependencies have
been already built and are the expected locations on the filesystem.

* Use modified cosmos-sdk
* Update Makefile
* Update go.mod and go.sum
* Add tx hash log
* Use local cosmos-sdk
* hack: context time